refactor(card): moved card from beta to stable
moved card from beta to stable

COMUI-1875
