--- packages/genesys-spark-components/src/components/beta/gux-card/gux-card.types.ts
@@ -1,3 +1,0 @@
-export type GuxCardAccent = 'outline' | 'filled' | 'raised';
-
-export type GuxCardAccentV4 = 'bordered' | 'filled' | 'raised' | 'borderless';
