+++ packages/genesys-spark-components/src/components/stable/gux-card/example.html
@@ -1,45 +1,29 @@
-<h1>gux-card-beta</h1>
+<h1>gux-card</h1>
 
 <div class="container">
-  <gux-card-beta accent="outline">
-    <h2>Outline</h2>
-    <p>
-      Lorem ipsum dolor sit amet, consectetur adipiscing elit. Nam convallis
-      erat fringilla tristique condimentums.
-    </p>
-  </gux-card-beta>
-
-  <gux-card-beta accent="bordered">
+  <gux-card accent="bordered">
     <h2>Bordered</h2>
     <p>
       Lorem ipsum dolor sit amet, consectetur adipiscing elit. Nam convallis
       erat fringilla tristique condimentums.
     </p>
-  </gux-card-beta>
+  </gux-card>
 
-  <gux-card-beta accent="filled">
-    <h2>Filled</h2>
-    <p>
-      Lorem ipsum dolor sit amet, consectetur adipiscing elit. Nam convallis
-      erat fringilla tristique condimentums.
-    </p>
-  </gux-card-beta>
-
-  <gux-card-beta accent="raised">
+  <gux-card accent="raised">
     <h2>Raised</h2>
     <p>
       Lorem ipsum dolor sit amet, consectetur adipiscing elit. Nam convallis
       erat fringilla tristique condimentums.
     </p>
-  </gux-card-beta>
+  </gux-card>
 
-  <gux-card-beta accent="borderless">
+  <gux-card accent="borderless">
     <h2>Borderless</h2>
     <p>
       Lorem ipsum dolor sit amet, consectetur adipiscing elit. Nam convallis
       erat fringilla tristique condimentums.
     </p>
-  </gux-card-beta>
+  </gux-card>
 </div>
 
 <style>
@@ -51,7 +35,7 @@
     background-color: #f2f2f2;
   }
 
-  gux-card-beta {
+  gux-card {
     width: 200px;
   }
 </style>
