+++ packages/genesys-spark-components/src/components/stable/gux-card/gux-card.types.ts
@@ -1,0 +1,1 @@
+export type GuxCardAccent = 'bordered' | 'raised' | 'borderless';
